+++ src/main/java/com/metric/oyometrics_team2/DTO/FirstEventResponse.java
@@ -13,5 +13,5 @@ import org.springframework.stereotype.Component;
 @ToString
 @Component
 public class FirstEventResponse {
-    private Long FirstReviewMeanTime;
+    private double FirstReviewMeanTime;
 }
